requirements.txt
discord~=2.3.2
asyncpg~=0.30.0
python-dotenv~=1.0.1
setuptools~=60.2.0